수정: Page/ Styles 분리
