+++ src/App.tsx
@@ -1,5 +1,3 @@
-console.log("This is Dev Branch!");
-
 import React from 'react';
 import { createGlobalStyle, ThemeProvider } from 'styled-components';
 import { useRecoilValue } from 'recoil';
